Fixed the bug

+++ packages/widgets/tenant-profile-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initTenantCustomAttributesMixin.ts
@@ -148,24 +148,30 @@ export const initTenantCustomAttributesMixin = createSingletonMixin(
       ) {
         const editFlowId = nodeEle.getAttribute('edit-flow-id');
         if (editFlowId) {
-          this.#editModals[editFlowId] = this.createModal({
-            'data-id': `edit-${customAttrName}`,
-          });
+          // Only create modal and flow if they don't exist yet
+          if (!this.#editModals[editFlowId]) {
+            this.#editModals[editFlowId] = this.createModal({
+              'data-id': `edit-${customAttrName}`,
+            });
 
-          this.#editFlows[editFlowId] = new FlowDriver(
-            () =>
-              this.#editModals[editFlowId]?.ele?.querySelector('descope-wc'),
-            { logger: this.logger },
-          );
-          this.#editModals[editFlowId].beforeOpen =
-            this.#initEditModalContent.bind(this, editFlowId);
+            this.#editFlows[editFlowId] = new FlowDriver(
+              () =>
+                this.#editModals[editFlowId]?.ele?.querySelector('descope-wc'),
+              { logger: this.logger },
+            );
+            this.#editModals[editFlowId].beforeOpen = () => {
+              const currentVal = getTenantCustomAttributes(this.state)[attName];
+              this.#initEditModalContent(editFlowId, type, attName, currentVal);
+            };
 
-          compInstance.onEditClick(() => {
-            this.#editModals?.[editFlowId]?.open();
-          });
+            compInstance.onEditClick(() => {
+              this.#editModals?.[editFlowId]?.open();
+            });
+
+            this.syncFlowTheme(this.#editFlows[editFlowId]);
+          }
 
           this.#initEditModalContent(editFlowId, type, attName, val);
-          this.syncFlowTheme(this.#editFlows[editFlowId]);
         }
       }
 
@@ -176,26 +182,30 @@ export const initTenantCustomAttributesMixin = createSingletonMixin(
       ) {
         const deleteFlowId = nodeEle.getAttribute('delete-flow-id');
         if (deleteFlowId) {
-          this.#deleteModals[deleteFlowId] = this.createModal({
-            'data-id': `delete-${customAttrName}`,
-          });
+          // Only create modal and flow if they don't exist yet
+          if (!this.#deleteModals[deleteFlowId]) {
+            this.#deleteModals[deleteFlowId] = this.createModal({
+              'data-id': `delete-${customAttrName}`,
+            });
 
-          this.#deleteFlows[deleteFlowId] = new FlowDriver(
-            () =>
-              this.#deleteModals[deleteFlowId]?.ele?.querySelector(
-                'descope-wc',
-              ),
-            { logger: this.logger },
-          );
-          this.#deleteModals[deleteFlowId].beforeOpen =
-            this.#initDeleteModalContent.bind(this, deleteFlowId);
+            this.#deleteFlows[deleteFlowId] = new FlowDriver(
+              () =>
+                this.#deleteModals[deleteFlowId]?.ele?.querySelector(
+                  'descope-wc',
+                ),
+              { logger: this.logger },
+            );
+            this.#deleteModals[deleteFlowId].beforeOpen =
+              this.#initDeleteModalContent.bind(this, deleteFlowId);
 
-          compInstance.onDeleteClick(() => {
-            this.#deleteModals?.[deleteFlowId]?.open();
-          });
+            compInstance.onDeleteClick(() => {
+              this.#deleteModals?.[deleteFlowId]?.open();
+            });
+
+            this.syncFlowTheme(this.#deleteFlows[deleteFlowId]);
+          }
 
           this.#initDeleteModalContent(deleteFlowId);
-          this.syncFlowTheme(this.#deleteFlows[deleteFlowId]);
         }
       }
